Add check-auth and check-parse lambdas for login
Check auth lambda is used to check if the user has a valid cookie,
and refreshes it if near expiration. Check auth also redirects user
to login page if cookie is not valid or is missing.

Parse auth lambda is used for auth flow with cognito, taking the
code from the url and exchanging it for user tokens, which is then
written to cookies.